ampdoc: Support config.urls updates if singleDoc
Set config.urls properties 'cdn' or 'geoApi' with meta tags when an AMP
page is viewed in singleDoc mode. For example,

  <meta name="amp-config-urls-cdn" content="/myruntime">

will rewrite config.urls property 'cdn' early enough in ampdoc creation
that all resources which would normally be fetched from
https://cdn.ampproject.org/ will instead be fetched from relative path
`/myruntime` on the same origin as the AMP page is being viewed.